Implement transaction logic within Usd Stage.
Ensure that transactions can be nested by keeping a list of transaction
handlers which can be merged with the next one until all notices can be
processed. At the end of a transaction, mergeable notices will be
consolidated so that only one will be sent.

Outside of transactions, notices must be sent directly to ensure
backward compatibility.