Added merge logic to stage notices
Changed the merge function to take in an r-value reference. This
forces callers to have to std::move the notice to be merged. This
further implies that the notice passed in to be merged may be
modified to make merging efficient.

Made LayerMutingChanged notice own its data instead of it storing
a reference. This was necessary since the merging logic modifies
the data stored in the notice. Also, we need to hold onto the data
in case of a transaction.

Implemented the merge for ObjectsChanged, and LayerMutingChanged
notices. The ObjectsChanged notice merges the resyncChanges map
and the infoChanges map between the two notices. Each entry in
the notice is std::move'd to the other notice.
The LayerMutingChanged notice merges the mutedlayers and
unmutedlayers vector between the two notices.